Add first draft try for blockwise hist
dask_histogram does not deal with chunked dimensions along which an
histogram is computed.
It rechunks so that it can feed the entire dimension at once.
Not good obviously because it can end with large chunks. Even if this is
embarrassingly parallel.

The way to deal with that (and that is used by xhistogram) is
dask.array.blockwise(). Then proceed to sum the correct dimensions.

This is a bit messy because we need to work with indices.

For some reason, the graph visualization of the histogram is not correct
when it comes to the time dimension (outer chunk, we loop over it). idk why.

We need to check that the histogram can be summed together.